changed LoadShaders.h to pull shaders from a file (file name doesn't matter).
Added a file2string function that will take a filename as input, malloc the size needed, and copy the file to the string (yes, it's in C, yes, I found it on a website, forgot the link).

added 2 shader files, .vert is vertex, .frag is fragment (filename doesn't matter, it's what i use on mine though).